Refactor customizations, add comments everywhere, customizations as a proxy to components in light theme to allow selective customizations in the theme not in Volto itself.
