+++ src/components/Navigation/Navigation.jsx
@@ -1,0 +1,199 @@
+// SemanticUI-free pre-@plone/components
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { compose } from 'redux';
+import { defineMessages, injectIntl } from 'react-intl';
+import cx from 'classnames';
+import { BodyClass, getBaseUrl, hasApiExpander } from '@plone/volto/helpers';
+import config from '@plone/volto/registry';
+import { getNavigation } from '@plone/volto/actions';
+import { CSSTransition } from 'react-transition-group';
+import NavItems from '@plone/volto/components/theme/Navigation/NavItems';
+
+const messages = defineMessages({
+  closeMobileMenu: {
+    id: 'Close menu',
+    defaultMessage: 'Close menu',
+  },
+  openMobileMenu: {
+    id: 'Open menu',
+    defaultMessage: 'Open menu',
+  },
+});
+
+/**
+ * Navigation container class.
+ * @class Navigation
+ * @extends Component
+ */
+class Navigation extends Component {
+  /**
+   * Property types.
+   * @property {Object} propTypes Property types.
+   * @static
+   */
+  static propTypes = {
+    getNavigation: PropTypes.func.isRequired,
+    pathname: PropTypes.string.isRequired,
+    items: PropTypes.arrayOf(
+      PropTypes.shape({
+        title: PropTypes.string,
+        url: PropTypes.string,
+      }),
+    ).isRequired,
+    lang: PropTypes.string.isRequired,
+  };
+
+  static defaultProps = {
+    token: null,
+  };
+
+  /**
+   * Constructor
+   * @method constructor
+   * @param {Object} props Component properties
+   * @constructs Navigation
+   */
+  constructor(props) {
+    super(props);
+    this.toggleMobileMenu = this.toggleMobileMenu.bind(this);
+    this.closeMobileMenu = this.closeMobileMenu.bind(this);
+    this.state = {
+      isMobileMenuOpen: false,
+    };
+  }
+
+  componentDidMount() {
+    const { settings } = config;
+    if (!hasApiExpander('navigation', getBaseUrl(this.props.pathname))) {
+      this.props.getNavigation(
+        getBaseUrl(this.props.pathname),
+        settings.navDepth,
+      );
+    }
+  }
+
+  /**
+   * Component will receive props
+   * @method componentWillReceiveProps
+   * @param {Object} nextProps Next properties
+   * @returns {undefined}
+   */
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    const { settings } = config;
+    if (
+      nextProps.pathname !== this.props.pathname ||
+      nextProps.token !== this.props.token
+    ) {
+      if (!hasApiExpander('navigation', getBaseUrl(this.props.pathname))) {
+        this.props.getNavigation(
+          getBaseUrl(nextProps.pathname),
+          settings.navDepth,
+        );
+      }
+    }
+  }
+
+  /**
+   * Toggle mobile menu's open state
+   * @method toggleMobileMenu
+   * @returns {undefined}
+   */
+  toggleMobileMenu() {
+    this.setState({ isMobileMenuOpen: !this.state.isMobileMenuOpen });
+  }
+
+  /**
+   * Close mobile menu
+   * @method closeMobileMenu
+   * @returns {undefined}
+   */
+  closeMobileMenu(e) {
+    if (!this.state.isMobileMenuOpen) {
+      return;
+    }
+    if (e.key && e.key !== 'Enter') {
+      return;
+    }
+    this.setState({ isMobileMenuOpen: false });
+  }
+
+  /**
+   * Render method.
+   * @method render
+   * @returns {string} Markup for the component.
+   */
+  render() {
+    return (
+      <nav className="navigation" id="navigation" aria-label="navigation">
+        <div className="hamburger-wrapper mobile tablet only">
+          <button
+            className={cx('hamburger hamburger--spin', {
+              'is-active': this.state.isMobileMenuOpen,
+            })}
+            aria-label={
+              this.state.isMobileMenuOpen
+                ? this.props.intl.formatMessage(messages.closeMobileMenu, {
+                    type: this.props.type,
+                  })
+                : this.props.intl.formatMessage(messages.openMobileMenu, {
+                    type: this.props.type,
+                  })
+            }
+            title={
+              this.state.isMobileMenuOpen
+                ? this.props.intl.formatMessage(messages.closeMobileMenu, {
+                    type: this.props.type,
+                  })
+                : this.props.intl.formatMessage(messages.openMobileMenu, {
+                    type: this.props.type,
+                  })
+            }
+            type="button"
+            onClick={this.toggleMobileMenu}
+          >
+            <span className="hamburger-box">
+              <span className="hamburger-inner" />
+            </span>
+          </button>
+        </div>
+        <div className="desktop menu computer large screen widescreen only">
+          <NavItems items={this.props.items} lang={this.props.lang} />
+        </div>
+        <CSSTransition
+          in={this.state.isMobileMenuOpen}
+          timeout={500}
+          classNames="mobile-menu"
+          unmountOnExit
+        >
+          <div key="mobile-menu-key" className="mobile-menu">
+            <BodyClass className="has-mobile-menu-open" />
+            <div className="mobile-menu-nav">
+              <div
+                role="button"
+                tabIndex={0}
+                onClick={this.closeMobileMenu}
+                onKeyUp={this.closeMobileMenu}
+              >
+                <NavItems items={this.props.items} lang={this.props.lang} />
+              </div>
+            </div>
+          </div>
+        </CSSTransition>
+      </nav>
+    );
+  }
+}
+
+export default compose(
+  injectIntl,
+  connect(
+    (state) => ({
+      token: state.userSession.token,
+      items: state.navigation.items,
+      lang: state.intl.locale,
+    }),
+    { getNavigation },
+  ),
+)(Navigation);
