+++ src/customizations/@root/theme.js
@@ -1,3 +1,5 @@
+// This customization allows to enable the scape hatch to use
+// SCSS as a theme to override the SemanticUI theme.
 import 'semantic-ui-less/semantic.less';
 import '@plone/volto/../theme/themes/pastanaga/extras/extras.less';
 import '@kitconcept/volto-light-theme/theme/main.scss';
