+++ src/customizations/volto/components/manage/Blocks/Search/components/SearchDetails.jsx
@@ -1,3 +1,7 @@
+/**
+ * OVERRIDE SearchDetails.jsx
+ * REASON: Special template for displaying this component, use h2 instead of h4
+ */
 import React from 'react';
 import { defineMessages, useIntl } from 'react-intl';
